refactor: move adb shell mocking in measure package
It's not super well done, but let's keep what we already had for Flipper for now, and ensure we can run the same kind of test
