+++ packages/measure/src/__tests__/utils/sample-command-output-1.txt
