+++ packages/measure/src/__tests__/utils/sample-gfxinfo-output.txt
